// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      sites: {
        Row: {
          city: string | null
          created_at: string | null
          employer_id: number | null
          post_code: string | null
          project_type: string | null
          site_id: number
          site_name: string | null
          state: string | null
          street_address: string | null
          supervisor_name: string | null
          supervisor_telephone: string | null
          updated_at: string | null
        }
        Insert: {
          city?: string | null
          created_at?: string | null
          employer_id?: number | null
          post_code?: string | null
          project_type?: string | null
          site_id?: number
          site_name?: string | null
          state?: string | null
          street_address?: string | null
          supervisor_name?: string | null
          supervisor_telephone?: string | null
          updated_at?: string | null
        }
        Update: {
          city?: string | null
          created_at?: string | null
          employer_id?: number | null
          post_code?: string | null
          project_type?: string | null
          site_id?: number
          site_name?: string | null
          state?: string | null
          street_address?: string | null
          supervisor_name?: string | null
          supervisor_telephone?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "workplace_locations_employer_id_fkey"
            columns: ["employer_id"]
            isOneToOne: false
            referencedRelation: "employers"
            referencedColumns: ["employer_id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never